% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/craninfo.R
\name{craninfo}
\alias{craninfo}
\title{FUNCTION_TITLE}
\usage{
craninfo(
  cran_type = "release",
  platform = get_os(),
  machine = R.version$arch,
  linux_type = ifelse(platform == "linux" & cran_type == "devel", "debian-clang",
    NA_character_)
)
}
\arguments{
\item{cran_type}{PARAM_DESCRIPTION, Default: 'release'}

\item{platform}{PARAM_DESCRIPTION, Default: get_os()}

\item{machine}{PARAM_DESCRIPTION, Default: R.version$arch}

\item{linux_type}{PARAM_DESCRIPTION, Default: ifelse(platform == "linux", "debian-clang", NA_character_)}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
