% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_os.R
\name{get_os}
\alias{get_os}
\title{FUNCTION_TITLE}
\usage{
get_os()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
